Attack/requirements.txt
pycryptodome==3.17
pyusb==1.2.1
scapy==2.5.0